[REFACTOR-018] extract orgchart-math, los buildTree, LosToolbar
